requirements.txt
agora_token_builder==1.0.0
Flask==2.2.5
Flask_Avatars==0.2.3
Flask_Login==0.6.3
Flask_Mail==0.9.1
flask_sqlalchemy==3.1.1
flask_wtf==1.2.1
python-dotenv==1.0.1
stripe==8.9.0
Werkzeug==3.0.2
WTForms==3.1.2
gunicorn
email_validator